requirements.txt
azure-ai-formrecognizer
azure-core
groq
python-dotenv